# cedar-worker-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe worker server runs the background jobs the rest of CEDAR cannot do inside a request: work\nthat takes longer than a caller should wait, and that must survive being interrupted.\n\nIts endpoints start a job and report on one. They are restricted to administrators.\n",
    "title" : "CEDAR Worker Server API",
    "version" : "2.9.7"
  },
  "paths" : {
    "/command/regenerate-inclusion-subgraph" : {
      "post" : {
        "description" : "Ask for the graph of which artifacts include which to be rebuilt, and return the job that will do it. The work is asynchronous, so this answers immediately with a job to poll rather than waiting for the rebuild. Only one job runs at a time: asking again while one is in flight returns that job rather than starting a second. Restricted to administrators.",
        "operationId" : "regenerateInclusionSubgraph",
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InclusionSubgraphJob"
                }
              }
            },
            "description" : "A job was started",
            "headers" : {
              "Location" : {
                "description" : "Where to poll this job's status.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "The caller is not an administrator"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InclusionSubgraphJob"
                }
              }
            },
            "description" : "A regeneration is already running; the job returned is that one, not a new one",
            "headers" : {
              "Location" : {
                "description" : "Where to poll the running job's status.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Start regenerating the inclusion subgraph",
        "tags" : [ "Inclusion subgraph" ]
      }
    },
    "/command/regenerate-inclusion-subgraph/{jobId}" : {
      "get" : {
        "description" : "Report where a regeneration job has got to. Jobs are held in memory, so an identifier from before the last restart is no longer known. Restricted to administrators.",
        "operationId" : "regenerationStatus",
        "parameters" : [ {
          "description" : "Job identifier, as returned when the job was started.",
          "in" : "path",
          "name" : "jobId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InclusionSubgraphJob"
                }
              }
            },
            "description" : "The job and its current state"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "The caller is not an administrator"
          },
          "404" : {
            "description" : "No job answers to this identifier"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get the status of a regeneration job",
        "tags" : [ "Inclusion subgraph" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "InclusionSubgraphJob" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "completedAt" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "error" : {
            "type" : "string",
            "nullable" : true
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "status" : {
            "type" : "string",
            "enum" : [ "QUEUED", "RUNNING", "SUCCEEDED", "FAILED" ]
          },
          "submittedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "id", "status", "submittedAt" ]
      }
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}